+++ examples/example1-1-intro.py
@@ -1,5 +1,4 @@
 #!/usr/bin/env python3
-# -*- coding: utf-8 -*-
 
 # Just the default agent that takes data from the standard input and writes as-is to the standard output
 # Note : a Sink expects a NGSI DataModel class as input. If it is not the case it simply writes the string representation of the its input