+++ examples/example3-2-httpserver-agent-upload-json.py
@@ -1,5 +1,4 @@
 #!/usr/bin/env python3
-# -*- coding: utf-8 -*-
 
 # This agent takes data from a PUSH source, build NGSI entities, eventually writes to a given Sink.
 # https://fiware-orion.readthedocs.io/en/2.0.0/user/walkthrough_apiv2/index.html#entity-creation